OBPIH-7911 Fixes after review
