+++ src/js/utils/receiving/omitBlankReceivingRows.js
@@ -1,3 +1,5 @@
+import _ from 'lodash';
+
 import { filterLineItemsState } from 'utils/receiving/receivingRowFilter';
 
 /**
@@ -7,7 +9,8 @@ const omitBlankReceivingRows = (lineItemsState) => filterLineItemsState(
   lineItemsState,
   new Set(
     Object.values(lineItemsState?.entities || {})
-      .filter((row) => row?.quantityReceiving != null)
+      // Every row is checked, toggle rows included - those carry no quantity field at all.
+      .filter((row) => !_.isNil(row?.quantityReceiving))
       .map((row) => row.shipmentItemId),
   ),
 );
